Fix all ESLint errors and configure no-unused-vars for motion
